Cleanup global structs
Since the various response classes are only used to assist with
unmarshaling and are never returned or even exported, I have inlined the
struct definitions to keep the global scope a little cleaner.

In addition to this, I am relying on the much cleaner NewDecoder.Decode
method instead of json.Unmarshal.